Move tests in a specific namespace
