+++ tests/Filters/PostEqualsFilter.php
@@ -1,14 +1,11 @@
 <?php
 
+namespace Kblais\QueryFilter\Tests\Filters;
+
 use Kblais\QueryFilter\QueryFilter;
 
-class PostTwoFilters extends QueryFilter
+class PostEqualsFilter extends QueryFilter
 {
-    public function title($value)
-    {
-        return $this->like('title', $value);
-    }
-
     public function category($value)
     {
         return $this->equals('category', $value);
