+++ tests/Filters/PostNoFilter.php
@@ -1,5 +1,7 @@
 <?php
 
+namespace Kblais\QueryFilter\Tests\Filters;
+
 use Kblais\QueryFilter\QueryFilter;
 
 class PostNoFilter extends QueryFilter
